+++ src/plugins/card/card.js
@@ -47,7 +47,7 @@ async function doCard(msg) {
 
   data.emoticons = emoticons;
 
-  render(msg, data, "genshin-card-8");
+  render(msg, data, "genshin-card");
 }
 
 export { doCard };
