--- src/views/genshin-card-8.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-cn">
-  <head>
-    <meta charset="UTF-8" />
-    <title>card-8</title>
-    <script src="../../data/js/vue3.global.prod.js"></script>
-    <link rel="preload" href="../../data/font/HYWenHei-85W.ttf" as="font" type="font/ttf" crossorigin />
-    <link rel="preload" href="../../data/font/YunXiJinShu.ttf" as="font" type="font/ttf" crossorigin />
-    <link rel="preload" href="../../data/font/azuki.ttf" as="font" type="font/ttf" crossorigin />
-    <link rel="preconnect" href="https://upload-bbs.mihoyo.com" crossorigin />
-    <link rel="preload" as="script" href="component/common/utils.js" crossorigin />
-    <link rel="preload" as="script" href="component/card-8/app.js" crossorigin />
-    <link rel="preload" as="script" href="component/card-8/cardComponents.js" crossorigin />
-    <link rel="stylesheet" href="./component/card-8/card-8.css" />
-    <link rel="preload" as="image" href="../../resources/Version2/module/user-base.png" type="image/png" />
-    <link rel="preload" as="image" href="../../resources/Version2/components/title-arrow.svg" type="image/svg+xml" />
-    <link rel="icon" href="data:," />
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module">
-      import App from "./component/card-8/app.js";
-
-      Vue.createApp(App).mount("#app");
-    </script>
-  </body>
-</html>
