+++ src/views/genshin-card.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="zh-cn">
+  <head>
+    <meta charset="UTF-8" />
+    <title>genshin-card</title>
+    <script src="../../data/js/vue3.global.js"></script>
+    <link rel="preload" href="../../data/font/HYWenHei-85W.ttf" as="font" type="font/ttf" crossorigin />
+    <link rel="preconnect" href="https://upload-bbs.mihoyo.com" crossorigin />
+    <link rel="preload" as="script" href="component/common/utils.js" crossorigin />
+    <link rel="preload" as="script" href="component/card/app.js" crossorigin />
+    <link rel="preload" as="script" href="component/card/cardComponents.js" crossorigin />
+    <link rel="stylesheet" href="./component/card/card.css" />
+    <link
+      rel="preload"
+      as="image"
+      href="../../resources/Version2/backgrounds/contour-background.svg"
+      type="image/svg+xml"
+    />
+    <link rel="icon" href="data:," />
+  </head>
+  <body>
+    <div id="app"></div>
+    <script type="module">
+      import App from "./component/card/app.js";
+
+      Vue.createApp(App).mount("#app");
+    </script>
+  </body>
+</html>
